feat: implement user-specific task retrieval and authorization checks for updates and deletions.

+++ backend/controllers/taskController.js
@@ -1,12 +1,12 @@
-
 const asyncHandler = require('express-async-handler')
 const Task = require('../models/taskModel')
+const User = require('../models/userModel')
 
 // Retrieve all tasks
 // GET /api/tasks
 // Private
 const getTask = asyncHandler(async (req, res) => {
-  const tasks = await Task.find()
+  const tasks = await Task.find({ user: req.user.id })
   res.status(200).json(tasks)
 })
 
@@ -20,6 +20,7 @@ const setTask = asyncHandler(async (req, res) => {
   }
   const task = await Task.create({
     text: req.body.text,
+    user: req.user.id,
   })
   res.status(201).json(task) // 201 = Created
 })
@@ -35,10 +36,15 @@ const updateTask = asyncHandler(async (req, res) => {
     throw new Error('Task not found')
   }
 
+  if (task.use.toString() !== req.user.id){
+    res.status(401)
+    throw new Error('Not authorised to update task')
+  }
+
   const updatedTask = await Task.findByIdAndUpdate(
     req.params.id,
     { text: req.body.text },
-    { new: true } // returns the updated version, not the old one
+    { new: true } // returns the updated version
   )
 
   res.status(200).json(updatedTask)
@@ -55,8 +61,12 @@ const deleteTask = asyncHandler(async (req, res) => {
     throw new Error('Task not found')
   }
 
-  await task.deleteOne()
+  if (task.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('Not authorised to delete this task')
+  }
 
+  await task.deleteOne()
   res.status(200).json({ id: req.params.id, message: 'Task deleted' })
 })
